Added working modal
Used my current version of the code including features I've been working on, so just ignore everything else.  I added a link named "Statistics" above the title on the Home page.  When clicked, it redirects to a page where the created modal automatically opens.  I also added a button on this page that is supposed to make the modal reappear, but haven't gotten it to work yet.